add style to search zone
