+++ frontend/src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import Header from "@components/Header";
-import SearchTrain from "@components/SearchTrain";
+import SearchTrain from "@components/SearchTrains";
 
 export default function Home() {
   return (
